+++ .pre-commit-config.yaml
@@ -1,0 +1,12 @@
+repos:
+-   repo: https://github.com/ambv/black
+    rev: 19.3b0
+    hooks:
+    - id: black
+-   repo: https://gitlab.com/PyCQA/flake8
+    rev: '3.7.8'
+    hooks:
+    - id: flake8
+      additional_dependencies: ['pep8-naming']
+    # Ignore all format-related checks as Black takes care of those.
+    args: ['--ignore', 'E2,W5', '--select', 'E,W,F,N', '--max-line-length=90']
