+++ geopatra/__init__.py
@@ -1,3 +1,5 @@
+# flake8: noqa
+
 __version__ = "0.1.1"
 
-from geopatra.map import InteraciveMap
+from .map import InteractiveMap
