+++ geopatra/map.py
@@ -1,31 +1,37 @@
 """Combine plots and expose them with geopandas."""
 
 
-import geopandas as gpd
-from .folium import geojson, chloropeth, markercluster, heatmap
+import geopandas
+from geopatra import folium
 
 
-@gpd.tools.util.pd.api.extensions.register_dataframe_accessor("folium")
-class InteraciveMap:
+@geopandas.tools.util.pd.api.extensions.register_dataframe_accessor("folium")
+class InteractiveMap:
+    """Extends geopandas to enable easy interactive plotting with folium.
+
+    Refer folium/ to call the below functions directly or to customize them
+    """
+
     def __init__(self, geopandas_obj):
+        """Init InteractiveMap with geodataframe."""
         self._gdf = geopandas_obj
 
     def plot(self, **kwargs):
-        gdf = self._gdf
-        m = geojson(gdf, **kwargs)
+        """Plot geodataframe as a geojson."""
+        m = folium.geojson(self._gdf, **kwargs)
         return m
 
     def chloropeth(self, **kwargs):
-        gdf = self._gdf
-        m = chloropeth(gdf, **kwargs)
+        """Plot geodataframe as a chloropeth."""
+        m = folium.chloropeth(self._gdf, **kwargs)
         return m
 
     def markercluster(self, **kwargs):
-        gdf = self._gdf
-        m = markercluster(gdf, **kwargs)
+        """Plot geodataframe as a markercluster."""
+        m = folium.markercluster(self._gdf, **kwargs)
         return m
 
     def heatmap(self, **kwargs):
-        gdf = self._gdf
-        m = heatmap(gdf, **kwargs)
+        """Plot geodataframe as a heatmap."""
+        m = folium.heatmap(self._gdf, **kwargs)
         return m
